+++ Package.swift
@@ -9,9 +9,15 @@ let package = Package(
     products: [
         .library(name: "ColorCode", targets: ["ColorCode"]),
     ],
+    dependencies: [
+        .package(url: "https://github.com/apple/swift-numerics", from: "1.0.0"),
+    ],
     targets: [
         .target(name: "ColorCode"),
-        .testTarget(name: "ColorCodeTests", dependencies: ["ColorCode"])
+        .testTarget(name: "ColorCodeTests", dependencies: [
+            "ColorCode",
+            .product(name: "Numerics", package: "swift-numerics"),
+        ])
     ],
     swiftLanguageVersions: [.v6]
 )